move search box to search folder
